fix(api): change trash definitions to avoid y head crash (#4188)
In 42deac230ce334e0a6ace69b756a1df5d8dd63e0 we changed the behavior of container
load to load v2 trash containers and dynamically modify the geometry to move the
"center" point +y by half the size of a multichannel pipette. This let our
multichannel pipettes dispense in the middle of the trash without having to
special case behavior at runtime. It also let us use the v2 trash definitions in
apiv1, which are more correct.

Unfortunately, one of the ways in which the old trash definitions were incorrect
was to put the "center" of the trash well at a very specific location for both
multi and single pipettes - y+80mm. With the correct geometry and special casing
based on the size of a multi, we were using y+85mm or so. This interacts badly
with a sheet metal change on the gantry, and robots using these definitions
collide when dropping tips in the trash.

This commit removes the geometry modifications at runtime in favor of altering
the v2 trash geometry to place the y-center at the exact right place. It also
removes the centerMultichannelOnWells quirk from the new trash defs since that
isn't accurate anymore - the backmost channel of the multi should go to this
very specific position.