refactor(api): Implement an API v1 Executor (#4176)
With this PR, you can now upload an api v1 protocol with the api v2 feature flag flipped. Note The protocol will not run correctly as the rest of the tickets in the backcompat milestone. This PR accomplishes the binding of the backcompat globals into importable names for protocols and provides structure for the rest of the backcompat execution.

In addition, tests are now run in two passes: one for apiv1, one for apiv2. This gets around the import-time behavior changes based on the feature flag.


Closes #3538