fix(api): apiv1: touch tip before blowing out during transfers
A couple years ago, transfer, distribute, and consolidate were changed to always
blow out in the trash. Unfortunately, touch tip remained _after_ the blowout.
This meant that when a blowout was involved, the robot would dispense, then move
to the trash, then blow out, then touch tip (somewhere, somehow) in the trash.
This might lead to drips from the pipette as it moved between the destination
and trash contaminating wells in its path.

This commit fixes the issue by always blowing out, which involves moving to the
trash, after touching tip.

Closes #419