refactor(api): make tempdeck set_temp in apiv2 behave like TC
In the interest of consistency, this makes the tempdeck's set_temp call hold until the temperature
is reached by default. This is the same behavior as the thermocycler api. The ability to continue
work while an on going set temperature command is occuring will be covered by an upcoming addition
that handles that concern at the top protocol scope.