refactor(api): add restart links to settings
POST /settings now has a "links" field in its response. If the "links" field
contains a "restart" element, that element is a URI pointing to /server/restart,
and its presence indicates that the client should restart the robot using that
API.

The restart link is present if a setting that has the "restart_required" field
is changed.