fix(api,shared-data,labware-creator): do not touch tip on troughs (#4271)
* refactor(shared-data): add touchTipDisabled quirk to troughs, trash

Protocols cannot specify touch tip to reservoirs and the trash bin, as it's
meaningless in the second case and produces very bad behavior in the first. Add
a quirk to these labware that disables touchtip.

* fix(api): apiv2: do not allow touch tip on reservoir or trash

On labware with the 'touchTipDisabled' quirk, do nothing in touch tip. This
matches the behavior of apiv1.

* fix(labware-creator): emit touchTipDisabled quirk for troughlike lw

We're adding the `touchTipDisabled` quirk to prevent the api from touching tip
on labware where it's not a good idea, like the trash or like troughs. This
commit adds that capability to labware creator.

Closes #4258