refactor(api): add restart links to settings (#4290)
POST /settings and GET /settings now have "links" field in their responses. If the "links" field
contains a "restart" element, that element is a URI pointing to /server/restart,
and its presence indicates that the client should restart the robot using that
API.

The restart link is present if a setting that has the "restart_required" field
is changed. This state is latching: if a setting that has the `restart_required` field set is changed, restart will be indicated until the robot restarts. This is because the field that currently requires this (the API2 feature flag) really needs to be changed once and then immediately restarted, because it has some bad effects on the running server.