refactor(api): make apiv2 tempdeck set_temp behave like TC (#4273)
In the interest of consistency, this makes the tempdeck's set_temp call hold until the temperature is reached by default. This is the same behavior as the thermocycler api. The ability to continue work while an on going set temperature command is occuring will be covered by an upcoming addition that handles that concern at the top protocol scope.