feat(api): reflect original instrument name via rpc
When the user specifies an instrument as the name ("p300_single", "p10_single")
but a compatible GEN2 pipette is connected, the instrument name in rpc shows up
as the pipette connected to the robot.

Add an RPC-visible element called "requested_as" to the instrument model that is
the string the user used to request the instrument.