fix(api): apiv2: initialize simulator runflag (#4330)
In
150c7847994962733cbcc786b50b315d78562d59 (https://github.com/Opentrons/opentrons/pull/4294)
we changed the sense of the run flag in the apiv2 simulator to be clear = pause,
set = run. However, we didn't initialize it to set, so we always get the error
message.

It's also unclear if this error message is useful. A pause without a
corresponding resume most likely means the user wants to pause and not resume
until they click resume in the run app, which is a normal case.

This initializes the flag but also removes the warnings.