feat(api): apiv2: Limit protocols to max supported version
When we create a protocol context, give it the requested protocol API version.
Do not allow the simulation or execution of protocols that request a version
above the highest version this release of the software supports.

This behavior should be consistent across all entry points: rpc protocol
session, opentrons_execute/get_protocol_api,
opentrons_simulate/get_protocol_api. The get_protocol_api functions now take an
optional argument to limit the protocol api version to user specifications (with
a default of the max supported version)

Closes #4342