refactor(api): apiv2: tag all methods and attrs with version added
This adds a do-nothing decorator that lets you specify the version of an API
call.

It also changes the module firmware uploaders in apiv2 to use a temporary config
file for avrdude instead of getting the path to the file in the wheel because
it's technically better practice and to resolve a circular import issue.

Closes #4339