chore(docs): apiv2: Document the version all ifaces were added
The requires_version decorator now adds a sphinx version-added role to the
docstring of whatever it decorates (with some odd spacing to make it parse
correctly). This causes a nice little "New in version x.y" line to show up in
the autogenerated docs.

In addition, manually add versionadded stanzas to all the docs that mention
specific methods or attributes.

Closes #4340