feat(api): apiv2: check protocol versions when using api
The requires_version decorator now checks that any decorated function was active
at the requested API level when it is called. If not, there's an informative
error.

More behavior may have to be added to this in the future to handle deprecations;
however, for now this will check minimum supported version.

Closes #4343