fix(api,shared-data): fix gen2 multi positioning
Our references for the height (and therefore homing position) of the gen2 multi
was not taking into account that gen1 pipettes home to a bearing, not to the top
of their housing; we have to modify the positions a little bit.

Also, the gen2 multis are tall enough that we need to bump the attach-tip
position for deck calibration down a bit to avoid triggering the z endstop.