feat(api): apiv2: backcompat: implement LegacyWell position methods
The old Placeable wells had different arguments in  their positioning methods
than the new Wells, and also returned a different thing - a named tuple with
arguments reversed compared to Location, and also with the point encoding an
_offset_ relative to the well origin in the bottom left rather than an absolute
point in deck coordinates.

The new LegacyLocation is meant to encode such an offset pairing (and has the
same ordering as the tuples customarily used in the Pipette and Robot classes),
and the accessors now have the right arguments and return the correct
LegacyLocations, as verified through comparison to v1 outputs.