fix(api): connect context to real thermocycler during calibration
The calibration session currently picks up the simulated context from the beginning of the session.
This means that the simulated modules are also usedduring labware calibration. This temporarily
connect the thermocycler during calibration so the session is aware of the lid's open/close state.
This is a short term patch that will be removed in place of better session management to replace the
current RPC over HTTP session communication.