feat(api): raise an error when uploading a v2 protocol on v1 server
In the general case, since V1 doesn't have explicit versioning information, we
want to allow people to upload whatever they want to a V1 server. However, if
somebody uploads a protocol that has its apiLevel key in the metadata set to 2.0
or higher, we can be pretty certain they just forgot (or didn't know) to flip
the v2 feature flag, so let's tell them about it.