refactor(api): make labware migration non-destructive (#4476)
Since we're going to spend a while with the ability to switch back and forth
between v1 and v2, we should make the v1->v2 labware migration not delete the
labware database when it's done. Instead, we just skip any labware we've already
migrated.

This also means we don't have to gate this functionality behind an environment
variable anymore and we can just do it all the time (even if we're on v1
internals).