fix(api): prevent crashes after downgrade from adv settings changes
Unfortunately, the API server (before this PR) crashes if it is missing things
in its advanced settings file that should be there given the version. That means
that if we delete the useProtocolApi2 keys and useApi1BackCompat keys, then if
the user downgrades away from this version their server won't boot. We have to
keep those keys around, even if we don't look at them anymore.

To prevent this from reoccuring in the future, in addition to migrating the
advanced settings we also go through and make sure we have an entry of some kind
for every settings key we're expecting.