refactor(api): always use legacy singletons to run v1 protocols
Instead of using a back compat shim, we can parse the protocol and determine
what api version it supports on upload; then, if it is v1, we can (possibly
just-in-time build and) use the APIv1 singletons to simulate, calibrate, and
execute it entirely separately from the normal hardware controller.

This approach has downsides:
- you have two separate entities communicating with shared hardware, both
smoothie and module
- you have a significant amount of duplicated code

But it does guarantee feature parity while the rest of the code is switched to
v2, and the possible conflicts are minimized by only using the legacy singletons
during protocol execution.