fix(api): present loaded but unused pipettes and modules to rpc (#4538)
We've had a consistent issue in the session management around protocols that the
required instruments, modules, and labware are all pulled from scraping the
commands emitted by the protocol. That means if you load a pipette, instrument,
or module (which does not emit a command notification) but never use it (which
does), it never gets added to the requisite rpc-exposed property and thus
the app doesn't let you calibrate or indicate that you should attach the
pipette.

In V1, this was acceptable because as a byproduct of the way v1 worked, when you
ran the protocol it would create a handler for that pipette or module with no
issue.

In V2 (and also now when interpreting JSON protocols, since that uses V2) this
isn't true - we're much more strict about only allowing Contexts for the
hardware actually attached to the robot, and you always get an error at runtime.

This is the rightish behavior - we should force users to attach the hardware
they specify in their protocol to the robot - but the wrong way to expose it:
the required pipettes and modules, whether or not they're used, should show up
in the protocol's info page so the client can make you attach the right thing.

This commit does that by adding the list of attached instruments and modules
from the simulating protocol context to the rpc-exposed properties.

If the protocol is a V1 python protocol, that context is unused and the behavior
will remain the same as before.