fix(api): extend homing ack timeouts to avoid homing twice
After a recent refactor the smoothie driver has a short timeout for sending
commands to be acked and a longer timeout after sending an M400 to wait for the
command to execute. This works for most long-running commands like move, but
some commands - specifically G28.2, home - are executed before they ack rather
than being queued, and can take a long time. That means that if the gantry is in
the front left of the robot, as far as it can be from home, the x home will
frequently hit its short ack timeout and retry, which leads to two successive x
homes. This can also happen with plunger homes, especially on gen2 pipettes,
when the plunger is at the bottom.