feat(api): Handle extra labware in script entrypoints (#4574)
Previously, extra_labware (the mechanism for v2 custom labware) wasn't really
handled in the script entrypoints opentrons_execute and opentrons_simulate. This
adds that handling. Specifically,

- opentrons_execute learned -L, which lets you specify paths to search for
labware definitions (previously only known by opentrons_simulate for building
bundles). This defaults to '.', as in opentrons_simulate
- opentrons.{simulate,execute}.get_protocol_api learned the extra_labware
argument, which should be a dict mapping labware names to labware definitions.
If this is not specified (as in, not as an empty dict), if this is a robot the
software will look in /var/lib/jupyter/notebooks/labware for custom labware.
- Add the jupyter lookup thing to docs. This is the canonical "right place" to
put custom labware defs you want to use in jupyter.
- Add some tests.