fix(api): do not allow robot.connect() in rpc protocols
If you have a robot.connect() in your protocol, and upload it through the
opentrons app, then the robot will move and execute the commands in the protocol
during simulation. It will also behave badly when run. Make the protocol raise
an error if it calls robot.connect when run through the rpc session.

Closes #4252