refactor(api): look for smoothie firmware in /usr/lib/firmware (#4577)
That is the standard path for runtime-loaded firmware for attached hardware in
linux. It's a better way to update the smoothie than packing it in the api
wheel. In the future, we may even want to use a separate systemd task to update
the smoothie. Requires https://github.com/Opentrons/buildroot/pull/57