fix(api): special case pause at end of protocol (#4603)
If the last command in a protocol is a pause (or not necessarily the last
command, but with nothing that will actually pause in between it and the end of
the protocol) then the protocol will finish and the session will broadcast a
"finished" state, but will issue a home at the end of the protocol which will...
be paused. The app doesn't know because the session says it's finished, so the
app will never either automatically resume or present the option to resume, and
the only way out is to restart your robot.

Instead, at the end of a `run`, if the protocol is in a paused state, just wait.
This lets the app see the state and present the appropriate action to the user.

If the user resumes, the wait ends and we go on to end of protocol behavior. If
the user cancels, the smoothie is hard reset and then unpaused, and the
end-of-protocol home fails as it should.

Closes #3238