fix(api): specify v_offset in v1 transfer touch_tip (#4602)
When transfers emit a touch_tip in v1, they just pass the keyword arg value as
the first positional parameter. It is, however, a number, and interpreted as a
voffset by touch_tip; but it also emits a warning, which you can see if you're
simulating. Just pass it as the v_offset keyword arg.

No behavior change other than the warning going away is expected here.

Closes #3703