fix(api): make sure the robot is connected before querying instrs
Since the robot singleton is now not the thing the rest of the api runs on, it
may not be connected the first time the system tries to detect attached pipettes
before a run. This means that requested pipettes get built as v1, which means
that their settings stand a good chance of being wrong - and that can lead to
failed drop tips.

Closes #4590