refactor(update-server): Make regex string raw.
This fixes one of the warnings caught by the test suite.  The regex string relied on the deprecated behavior of unrecognized backslash escape sequences expanding to the literal backslash plus the literal character.  It's said that these will be syntax errors in the future.  Fix it by making the string raw.