fix(app): home all axes after lw calibration to allow deck access (#4687)
After returning the tip used in lw calibration, the gantry remains above the tiprack. This makes
accessing the items on the deck difficult if need be before the run. It is especially confusing in
the case of a TC protocol, as the user is prompted to place the PCR seal on the module, while the gantry may be physically blocking it. This appends the behavior of the final "return tip" command of lw calibration to include a "home all axes" command as well.

Closes #4034