feat(api): Add an OpenAPI spec for the HTTP API (#4691)
Add an OpenAPI spec (https://swagger.io/docs/specification/about/) to the robot HTTP API. This is a json document that lives in the source tree and is available from a live robot at GET /openapi. It describes the HTTP API of the robot. It will be kept up to date, checked at lint and test time, and used for client integration tests.

It also adds a schema validator and actually validates the schema, both in static
file form during the lint makefile task and as a test. It also fixes all the
problems with the openapi spec that the swagger visualizer didn't see

Closes #4635 

Co-authored-by: Mike Cousins <mike@cousins.io>