fix(api): set default move timeout to 12000 instead of 60 (#4768)
We added a split between ack and execute timeouts, with the execute timeout
being 12000 seconds. But we didn't actually change the places where the timeout
was specified to 30 for some reason. We therefore fixed the timeout issues that
consistently happened when the robot homed with the head far from its home...
but not the ones that happened when users specified a sharply limited speed.

Closes #4755