refactor(api): Add move splitting capability to driver move
This is in service of unsticking frozen plunger axes.

When multi pipettes of certain models sit idle for too long, their plunger o
rings create more friction the next time they try to move. To move the o rings,
we have to move a short distance slowly with a higher current before continuing
the rest of the motion.

While this could happen at a higher level, implementing this as splitting up a
move commanded for other reasons lets us more easily share the logic between the
v1 and v2 hardware handling code. Now, the hardware handling code has to check
how long since the plunger last moved and command the split if necessary (or, if
the parameters for the split end up meaning that not that much time elapses
during the slow move, just do it every time in some contexts).

Closes #4674