refactor(api): Add move splitting capability to driver move (#4848)
This is in service of unsticking frozen plunger axes.

When multi pipettes of certain models sit idle for too long, their plunger o
rings create more friction the next time they try to move. To move the o rings,
we have to move a short distance slowly with a higher current before continuing
the rest of the motion.

While this could happen at a higher level, implementing this as splitting up a
move commanded for other reasons lets us more easily share the logic between the
v1 and v2 hardware handling code. Additionally, move timing is reliant on checking the last motion _after_ any possible return from pause. That happens in the smoothie driver, so it has to be checked in the smoothie driver.

Closes #4674