fix(api) test_settings_endpoints fails when running independently.
Specifically, they pass the first time, but fails the second time. the root cause is that a robot settings file is created on the first run and breaks the second run. 

fixed by moving the mock_config to conftest and using it in the tests that require the config file teardown.

closes #4894