feat(api): support gen2 modules
Gen2 modules (with revisions above 20) are now detected as such. We use
the V2 module defs (if the protocol API version is high enough) to load
the modules, respecting the compatibility elements in the definitions.
We respect and apply the transformations.

This also adds the new moduleModel key to GET /modules and the model
element to the RPC module model.

Closes #4960