fix(api): hold motion lock in session for calling stop()
This prevents a race condition where the halt() and stop() calls could
happen in between the protocol run's smoothie calls, meaning that the
smoothie error response would be caught by the reset command instead of
the protocol. The subsequent command would just look like a warning, so
it wouldn't bubble up and the protocol wouldn't be cancelled.

Closes #4979