fix(api): swallow hard-halt-related errors and hold lock in session.stop() (#4980)
* fix(api): hold motion lock in session for calling stop()

This prevents a race condition where the halt() and stop() calls could
happen in between the protocol run's smoothie calls, meaning that the
smoothie error response would be caught by the reset command instead of
the protocol. The subsequent command would just look like a warning, so
it wouldn't bubble up and the protocol wouldn't be cancelled.

* further fix: don't swallow alarms

This could have been affecting other pause cases. There's many times in
the system where we wrap motion commands in try...finally to clean up.
Those cleanups usually involve sending other commands.

If you just did a smoothie hard halt, then the smoothie will respond to
those commands with "alarm lock" until you send an m999.

These will raise SmoothieErrors. Since this is in an exception context
it will put a new exception on the exception stack. That new exception
will be a smoothie error, and _that_ may be handled by code like fast
home and the cancel may never happen.

Now, we actually check to see if the error is an alarm lock or the
warning about homing and don't raise an error for them - there's already
been an error, the original kill button pressed one.

Closes #4979