fix(api): make probe failures raise the actual error
After c42d4dc62 (#4980), smoothie responses with downstream errors of halts (like the warnings
about having to home, or the "alarm lock" status condition) are
swallowed to make sure that the actual hard halt fault condition is
raised to the appropriate level, which makes it possible to separate
cancel behavior from actual error handling.

This coincidentally swallowed the error that happened when the user runs
a tip probe without a tip and the probe times out - even though there's
a special error code. That's because the tip probe execution timeout was
too low, and so the driver would time out and never read the actual
failure and only get the downstream errors (which would then be
swallowed).

The fix is to lengthen the timeout of the tip probe so the actual error
is raised.

While we're at it, make the error message human readable (and not
contain the word "smoothie")

Closes #5008