fix(api): do not check used tips after pickup
In api version 2.2, we stopped re-adding tips to the tip tracker after
return_tip to prevent subsequent automatic tip pickup (like calling
pick_up_tip without arguments, or the behavior in
transfer/consolidate/distribute calls) from re-using the tip that might
now be dirty.

However, at the end of pick_up_tip, the system tries to mark tips as
used and raises an error if any of the tips to be marked have been used.

The fix is to remove this check, since it is only relevant in the case
where pick_up_tip is called with a specific location argument, and in
that case we should assume the user really wants the behavior of reusing
the tip.

The root cause is that our tip tracker is about tracking tip presence,
not tip dirtiness. The behavior fixed in 2.2 was because of this: we
were only caring that the tip was physically back in the rack, so we
could reuse it. We now overload the tip tracker's results with tip
dirtiness and tip presence, and that needs to change.

Until that time, this fix should ensure the proper behavior.

Closes #5059