refactor(api): create a method that will initialize api.(#5124)
a slight refactor of main to create a method that will initialize the api. the purpose is to provide a method for the robot_server to initialize and use the opentrons api.  

in the future, api.main and api.server.main will be removed. all http server business will happen in robot_server.