fix(api): make api coroutine properties sync for polls during pause (#5166)
The hardware api had a couple of coroutine properties that wanted to be awaited inside of endpoints that are constantly polled during the run of a protocol.  With a central hardware thread, a pause will lead to the smoothie driver blocking the thread and sometimes causing the loop to time out. For now we can prevent this edge condition from crashing the session by ensuring polled endpoints only access sync properties and don't call into the hw event loop.

Co-authored-by: Seth Foster <seth@opentrons.com>