chore: add risk assessment field to pr template (#5233)
In a risk assessment, you go over your change set and identify other
areas of the system it might touch. Particularly important in a broad
and connected system, this process both gets the person submitting the
PR to consider the impact of their changes and points testers and
reviewers in the right direction.