fix(api): properly parse escaped nmcli responses
nmcli uses ':' as a field separator. If it encounters a : in a field
value, it escapes it with a backslash. We weren't parsing this
correctly (or indeed at all) and therefore couldn't connect to wifi
networks whose names include a colon. Parsing escaped strings is hard,
so this change uses the csv module to parse them properly