fix(api): make sure we move plunger to bottom after home
This commit make a new function _do_plunger_home which homes a plunger
on a mount, if it exists, and then moves that plunger to its configured
bottom. This is now used both by API.home_plunger (which was where this
behavior was before) but also by a call to API.home() that involves
pipettes, ensuring that plungers are always moved to the bottom after
they are homed.

Closes #5164