fix(api): reset location cache before tip probe
When tip probes fail, subsequent retries would see the robot lowering
the pipette dangerously to move back to the front of the robot for a new
tip. This was a combination of two issues:
- The location cache associated with the protocol context needed to be
cleared
- The smoothie driver automatically homes after errors like tip probe
failures, but doesn't tell the hardware controller, so it needs to
forcibly request the new position after a failure.

Closes #4793