refactor(build): add robot-server to top level push target
You can now do `make push` in the root of the repo and it will push the
robot server as well as the api and update server. previously it would
only do api and update, and you'd have to do a separate push to get the
robot server on there.