fix(api): exec protocols in module mode
if you give exec() separate locals and globals, it executes the provided
data as if embedded in a class definition which we very much do not want
- we want it to be executed as if in just a file, which of course it is.
This makes scoping work the same as it does when people write their
protocols and expect them to work.

While we're at it, call the argument we feed to the protocol's run
method something a little less likely to collide.

Closes #4982
Closes #4981