chore(api): enable typechecking of all defs (#5272)
By default, mypy will not typecheck functions that do not have type
annotations in their signatures. This is meant to aid in gradually
introducing typechecking to a pre-existing codebase. In a codebase like
ours where the expectation is that everything will be typechecked,
however, it just leads to unexpected failures to typecheck when you
forget to annotate functions (usually functions that have no arguments,
like __init__ or __iter__ magic methods).

Enable checking of unannotated defs, except for legacy code that is soon
to be deleted, and fix all the issues this config change finds.