refactor(update-server): remove legacy impls (#5407)
Remove the balena and migration versions of the update server, which
have not been used in now almost a year. Everything has been stable and
we don't use them anyway.

In addition, bump the python version for the dev environment to python 3.7.

This also makes the windows CI no longer run tests.