fix(api): correctly set magdeck offsets (#5416)
In
https://github.com/Opentrons/opentrons/pull/5330/ (f1cfb967c18fd2f43ea962cddc86e10e95e1d9f3)
we added per-model offsets for the height from base command for
magdecks. Previously, this value had been set to 5. The PR incorrectly
uses 5mm true for gen2s, and doubles it for the incorrect units for
gen1. Instead, what we need to do is use 5 (bad) mm for gen1, and 2.5 mm
true for gen2.