push smoothie programming down to allow gpios
Smoothie programming requires gpios (to manipulate the rest and isp pins
of the board), and that is only available now when a gpio module is
built. That means we need to have the gpio module available for
programming, so we push it down into the hardware controller to make
sure it's available.

That also requires some gymnastics to get programming to work in the
hardware controller thread - namely, you have to call an annoying
undocumented method of asyncio in the main thread to get signal listeners
set up if the first subprocess call is in a worker thread; and the main
thread has to actually be running to react to the signal handlers that
let it know the subprocess has ended.