refactor(api): load v1 singletons lazily (#5413)
One of the most important parts of supporting APIv1 is continuing to
support direct execution of protocols as scripts, protocols that look
like this:

from opentrons import robot
robot.connect()

That means that the robot singleton has to be either created in the
package init, or has to be available by normal import means through
opentrons.robot. We don't want the second because we want to sequester
apiv1 in its own little subdirectory, so we went with the first. The
problem with the first, though, is that you have to build the singleton
on any import of anything inside the opentrons package, and that's
really slow especially on a raspberry pi.

However, python 3.7 includes
[pep-562](https://www.python.org/dev/peps/pep-0562/), which allows for a
__getattr__ at module scope that behaves like its object cousin. Using
this, we can defer singleton initialization until the singleton is
actually accessed. That means that importing the top level opentrons
package becomes much faster if you don't need the robot singleton, which
is an all around improvement.

We could take this a step farther if we wanted, and only import
opentrons.robot from opentrons.api.session if somebody actually uploads
a v1 protocol. That would fully defer loading the singleton until
needed, but it would make the first v1 protocol upload after boot take
an extra 30 seconds or so.