fix(api): verify the structure of v2 protocol ast (#5454)
V2 protocols have to have a basic minimum structure: a single run
function that can be called with a single positional argument, e.g.
run(ctx). We have been checking this in
protocol_api.execute._runfunc_ok, on the exec'd protocol by using the
inspect module and looking at the exec'd globals.

The problem with this is if somebody uploads a protocol that doesn't
match this, we won't catch it until after we exec the protocol. We do
allow protocols that do work at module scope (e.g. script-style), but
that's also what v1 protocols look like. So if someone had a v1 protocol
that they had mistakenly or absent-mindedly told the system was a v2
protocol in the metadata, like this:

from opentrons import robot

metadata = {"apiLevel": "2.3"}

robot.comment("Hello, world!")

then that protocol will execute during the exec, and you won't get a
good error.

The way to get around this is to verify the structure from the AST,
before we call exec(). That AST is parsed in protocols.parse. So we move
the basic structure checking there, looking for a function definition
with the right name.

This keeps the argument checking in the protocol_api.execute module
though for reasons of excessive complexity: it's quite a bit simpler to
check the structure of a function's arguments using the inspect module
on a live object than it is to parse the wonderful glory of python's
function signature AST with its four or five separate kinds of argument.