fix(api): home only the axis probing after error
Also, raise all errors and not just ones that match "probe failure" to
the user. This fixes an issue where the default home, which hits all
axes, doesn't know about the required currents for the plunger (since
they're handled elsewhere) and fails; this raised an error that didn't
talk about tip probe, so it got swallowed.