fix(api): dont join poller in __del__
This fixes an interesting edge case brought about by doing something
that is in retrospect wrong with the temperature module poller thread.

When you start a thread with daemon=True, then any threads that thread
creates will inherit daemon=True. We make the threadmanager threads
daemonic. That means the poller threads are typically daemonic. This is
usually fine!

However, when you are creating them in a short lived program - like
opentrons_simulate - it's not fine. Python programs end when all
non-daemonic threads are done. The remaining daemonic threads are
terminated, at some point during the interpreter shutdown, messily.

join() doesn't work by magic - it has internal locking around the
thread's running state. When the interpreter kills a daemonic thread, it
doesn't necessarily handle that locking cleanly, because it doesn't have
to - it's a daemonic thread that's getting killed.

So if you have a daemonic thread remaining after program execution -
which the poller is because it's created from ThreadManager's thread,
which is daemonic, and that's inherited - and if you try to join that
thread during some other execution context that happens sometime during
interpreter shutdown, like the cyclic garbage collector cleaning up and
triggering object __del__ methods, then you could get a race between
join() in the __del__ and the interpreter killing the daemonic thread,
and you lock up.

The solution: don't do that! Separate the "set the quit flag" and
"join" concepts in the poller; only join() when you actually need to -
when you need to wait until the poller thread exits before doing
something else - and in the case of __del__, just set the flag (in case
we're not daemonic) and let the thread end naturally.

Closes #5561